requirements.txt
snowflake-connector-python>=3.0.0
snowflake-snowpark-python>=1.11.0
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.18.0
altair>=5.0.0
